% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Ellis.Wrist.RF.HMM.R
\docType{data}
\name{Ellis.Wrist.RF.HMM}
\alias{Ellis.Wrist.RF.HMM}
\title{Ellis Wrist Classifier}
\format{
A list containing the HMM and the RF models.
\describe{
\item{hmm}{The HMM model}
\item{rf}{The RF model of class randomForest}
}
}
\source{
Ellis et al.Med Sci Sports Exerc. 2016 May;48(5):933-40.
}
\usage{
Ellis.Wrist.RF.HMM
}
\description{
HMM and RF models for applying the Ellis Wrist Classifier in the actimetric package.
}
\examples{
data(Ellis.Wrist.RF.HMM)
}
\keyword{datasets}
