% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LVAY.RF.Hip10.R
\docType{data}
\name{LVAY.RF.Hip10}
\alias{LVAY.RF.Hip10}
\title{School Age Hip Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{LVAY.RF.Hip10}{The RF model}
}
}
\source{
Trost SG, et al. Physiol Meas. 2014. PMID: 25340887
}
\usage{
LVAY.RF.Hip10
}
\description{
RF model for applying the School Age Hip Classifier in the actimetric package.
}
\examples{
data(LVAY.RF.Hip10)
}
\keyword{datasets}
