% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LVAY.RF.Wrist5.1.R
\docType{data}
\name{LVAY.RF.Wrist5.1}
\alias{LVAY.RF.Wrist5.1}
\title{School Age Wrist Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{LVAY.RF.Wrist5.1}{The RF model}
}
}
\source{
Trost SG, et al. Physiol Meas. 2014. PMID: 25340887
}
\usage{
LVAY.RF.Wrist5.1
}
\description{
RF model for applying the School Age Wrist Classifier in the actimetric package.
}
\examples{
data(LVAY.RF.Wrist5.1)
}
\keyword{datasets}
