% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preschool_hipfl_15s.R
\docType{data}
\name{preschool_hipfl_15s}
\alias{preschool_hipfl_15s}
\title{Preschool Hip Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{preschool_hipfl_15s}{The RF model}
}
}
\source{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\usage{
preschool_hipfl_15s
}
\description{
RF model for applying the Preschool Hip Classifier in the actimetric package.
}
\examples{
data(preschool_hipfl_15s)
}
\references{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\keyword{datasets}
