% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preschooltest.R
\docType{data}
\name{preschooltest}
\alias{preschooltest}
\title{Preschool Wrist Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{preschooltest}{The RF model}
}
}
\source{
Trost SG, et al. Med Sci Sports Exerc. 2018 Mar;50(3):634-641.
}
\usage{
preschooltest
}
\description{
RF model for applying the Preschool Wrist Classifier in the actimetric package.
}
\examples{
data(preschooltest)
}
\keyword{datasets}
