% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PS.RF.FL.Hip15.LagLead.R
\docType{data}
\name{PS.RF.FL.Hip15.LagLead}
\alias{PS.RF.FL.Hip15.LagLead}
\title{Preschool Hip Lag-lead Classifier}
\format{
A randomForest object to make predictions on new data
\describe{
\item{PS.RF.FL.Hip15.LagLead}{The RF model}
}
}
\source{
Trost SG, et al. Med Sci Sports Exerc. 2018 Mar;50(3):634-641.
}
\usage{
PS.RF.FL.Hip15.LagLead
}
\description{
RF model for applying the Preschool Hip Lag-lead Classifier in the actimetric package.
}
\examples{
data(PS.RF.FL.Hip15.LagLead)
}
\keyword{datasets}
