% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trostRF_7112014.R
\docType{data}
\name{trostRF_7112014}
\alias{trostRF_7112014}
\title{Adult Wrist Classifier (Trost)}
\format{
A randomForest object to make predictions on new data
\describe{
\item{trostRF_7112014}{The RF model}
}
}
\source{
Trost SG, et al. Med Sci Sports Exerc. 2018 Mar;50(3):634-641.
}
\usage{
trostRF_7112014
}
\description{
RF model for applying the Adult Wrist Classifier (Trost) in the actimetric package.
}
\examples{
data(trostRF_7112014)
}
\keyword{datasets}
