% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LVAY.RF.Hip10.R
\docType{data}
\name{LVAY.RF.Hip10}
\alias{LVAY.RF.Hip10}
\title{School Age Hip Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{LVAY.RF.Hip10}{The RF model}
}
}
\source{
Ahmadi MN, Pfeiffer KA, Trost SG.
Physical Activity Classification in Youth Using Raw Accelerometer Data from
the Hip. Meas Phys Edu Exerc Sci. 2020; 24(2):129-136. doi: 10.1080/1091367X.2020.1716768
}
\usage{
LVAY.RF.Hip10
}
\description{
RF model for applying the School Age Hip Classifier in the actimetric package.
}
\examples{
data(LVAY.RF.Hip10)
}
\references{
Ahmadi MN, Pfeiffer KA, Trost SG.
Physical Activity Classification in Youth Using Raw Accelerometer Data from
the Hip. Meas Phys Edu Exerc Sci. 2020; 24(2):129-136. doi: 10.1080/1091367X.2020.1716768
}
\keyword{datasets}
