% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LVAY.RF.Wrist5.1.R
\docType{data}
\name{LVAY.RF.Wrist5.1}
\alias{LVAY.RF.Wrist5.1}
\title{School Age Wrist Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{LVAY.RF.Wrist5.1}{The RF model}
}
}
\source{
Chowdhury AK, Tjondronegoro D,
Chandran V, Trost SG. Ensemble Methods for Classification of Physical Activities
from Wrist Accelerometry. Med Sci Sports Exerc. 2017 Sep;49(9):1965-1973.
doi: 10.1249/MSS.0000000000001291
}
\usage{
LVAY.RF.Wrist5.1
}
\description{
RF model for applying the School Age Wrist Classifier in the actimetric package.
}
\examples{
data(LVAY.RF.Wrist5.1)
}
\references{
Chowdhury AK, Tjondronegoro D,
Chandran V, Trost SG. Ensemble Methods for Classification of Physical Activities
from Wrist Accelerometry. Med Sci Sports Exerc. 2017 Sep;49(9):1965-1973.
doi: 10.1249/MSS.0000000000001291
}
\keyword{datasets}
