% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preschool_hipfl_15s.R
\docType{data}
\name{preschool_hipfl_15s}
\alias{preschool_hipfl_15s}
\title{Preschool Hip Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{preschool_hipfl_15s}{The RF model}
}
}
\source{
Trost SG, et al. Med Sci Sports Exerc. 2018 Mar;50(3):634-641.
}
\usage{
preschool_hipfl_15s
}
\description{
RF model for applying the Preschool Hip Classifier in the actimetric package.
}
\examples{
data(preschool_hipfl_15s)
}
\keyword{datasets}
