% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preschooltest.R
\docType{data}
\name{preschooltest}
\alias{preschooltest}
\title{Preschool Wrist Classifier}
\format{
A train.formula object to make predictions on new data
\describe{
\item{preschooltest}{The RF model}
}
}
\source{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\usage{
preschooltest
}
\description{
RF model for applying the Preschool Wrist Classifier in the actimetric package.
}
\examples{
data(preschooltest)
}
\references{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\keyword{datasets}
