% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PS.RF.FL.Wrist15.LagLead.R
\docType{data}
\name{PS.RF.FL.Wrist15.LagLead}
\alias{PS.RF.FL.Wrist15.LagLead}
\title{Preschool Wrist Lag-lead Classifier}
\format{
A randomForest object to make predictions on new data
\describe{
\item{PS.RF.FL.Wrist15.LagLead}{The RF model}
}
}
\source{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\usage{
PS.RF.FL.Wrist15.LagLead
}
\description{
RF model for applying the Preschool Wrist Lag-lead Classifier in the actimetric package.
}
\examples{
data(PS.RF.FL.Wrist15.LagLead)
}
\references{
Ahmadi MN, Pavey TG, Trost SG. Machine Learning
Models for Classifying Physical Activity in Free-Living Preschool Children.
Sensors (Basel). 2020 Aug 5;20(16):4364. doi: 10.3390/s20164364
}
\keyword{datasets}
