% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trostRF_7112014.R
\docType{data}
\name{trostRF_7112014}
\alias{trostRF_7112014}
\title{Adult Wrist Classifier (Trost)}
\format{
A randomForest object to make predictions on new data
\describe{
\item{trostRF_7112014}{The RF model}
}
}
\source{
Pavey TG, Gilson ND, Gomersall SR, Clark B,
Trost SG. Field evaluation of a random forest activity classifier for wrist-worn
accelerometer data. J Sci Med Sport. 2017 Jan;20(1):75-80.
doi: 10.1016/j.jsams.2016.06.003
}
\usage{
trostRF_7112014
}
\description{
RF model for applying the Adult Wrist Classifier (Trost) in the actimetric package.
}
\examples{
data(trostRF_7112014)
}
\references{
Pavey TG, Gilson ND, Gomersall SR, Clark B,
Trost SG. Field evaluation of a random forest activity classifier for wrist-worn
accelerometer data. J Sci Med Sport. 2017 Jan;20(1):75-80.
doi: 10.1016/j.jsams.2016.06.003
}
\keyword{datasets}
